requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.2
botocore==1.27.2
dj-database-url==0.5.0
Django==4.0.5
django-filter==21.1
django-storages==1.12.3
gunicorn==20.1.0
jmespath==1.0.0
Pillow==9.1.1
psycopg2==2.9.3
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
matplotlib==3.4.3

